chore(init): download files only if not exists

+++ neovim/init.sh
@@ -1,6 +1,10 @@
+#!/usr/bin/bash
+
 source ../utils.sh
 
-wget -O autoload/plug.vim https://github.com/junegunn/vim-plug/raw/0.11.0/plug.vim
+if [[ ! -f autoload/plug.vim ]]; then
+  wget -O autoload/plug.vim https://github.com/junegunn/vim-plug/raw/0.11.0/plug.vim
+fi
 
 create_folder_if_not_exists ../local/neovim/
 create_file_if_not_exists ../local/neovim/init.vim
